Simplify MIDI interrupts.
Interrupts to call MIDI functions were doing a lot of work, especially
MIDI write. Rather than doing the heavy lifting in the interrupt, the
interrupt just sets a flag that a MIDI write or read should be done, and
the main loop will do that at next possible opportunity and cancel the
flag.