Add BOOTDELAY config variable
By default, this is not set. If set, it will delay startup by that many
milliseconds. The use-case is for something like a "Sweet Sixteen"
device, which is _inside_ a Eurorack case, connected to an ER-301 over
I2C. The ER301 must boot first before I2C connection is attempted. When
the devices are hardwired inside a case, this is a simple way to confirm
that the 16n only connects after 301 bootup.

For most desktop 16n users, this is irrelevant, but it makes compiling
a "Sweet-Sixteen" compatible firmware straightforward now.