Support builds with reversed faders
This separates the meaning of "flip" into its two components:

- flipOrder - which fader is 1 and which is 16 etc.
- reverseMinMax - which end of the fader is which

By default, these are both the same, following the layout of the
original 16n in which the fader is at its lowest voltage at the bottom.

However, in order to support the Sweet Sixteen, we can define
REVERSED_FADERS, and the normal (non-flipped) order will be with the
highest voltage at the bottom. This is implemented by interpreting
reverseMinMax as the opposite of flip.

Or, in tabular form:

    | variable      | 16n   | S16   |
    +---------------+---+---+---+---+
    | flip          | 0 | 1 | 0 | 1 |
    | flipOrder     | 0 | 1 | 0 | 1 |
    | reverseMinMax | 0 | 1 | 1 | 0 |

This allows the Sweet Sixteen to be operated upside down, if desired.